+++ tests/testthat/test-hide-variables.R
@@ -42,81 +42,98 @@ with_mock_crunch({
         })
     })
 
-    vg <- VariableGroup("A group", entities = ds[c("gender", "birthyr")])
-
-    test_that("hideVariables with various input types makes the right request", {
+    hide_gender <- paste0(
+        'https://app.crunch.io/api/datasets/3/folders/hidden/',
+        ' ',
+        '{',
+            '"element":"shoji:catalog",',
+            '"index":{',
+                '"https://app.crunch.io/api/datasets/3/variables/gender/":{}',
+            '}',
+        '}'
+    )
+    hide_both <- paste0(
+        'https://app.crunch.io/api/datasets/3/folders/hidden/',
+        ' ',
+        '{',
+            '"element":"shoji:catalog",',
+            '"index":{',
+                '"https://app.crunch.io/api/datasets/3/variables/gender/":{},',
+                '"https://app.crunch.io/api/datasets/3/variables/starttime/":{}',
+            '},',
+            '"graph":[',
+                '"https://app.crunch.io/api/datasets/3/variables/birthyr/",',
+                '"https://app.crunch.io/api/datasets/3/variables/gender/",',
+                '"https://app.crunch.io/api/datasets/3/variables/starttime/"',
+            ']',
+        '}'
+    )
+    test_that("hideVariables with one variable to hide", {
         expect_PATCH(
             ds <- hideVariables(ds, "gender"),
-            "https://app.crunch.io/api/datasets/3/variables/",
-            '{"element":"shoji:catalog","index":{"https://app.crunch.io/api/datasets/3/variables/gender/":{"discarded":true}}}'
+            hide_gender
         )
-        ## Duplicates are handled
+    })
+    test_that("hideVariables handles duplicated inputs", {
         expect_PATCH(
             ds <- hideVariables(ds, c("gender", "gender")),
-            "https://app.crunch.io/api/datasets/3/variables/",
-            '{"element":"shoji:catalog","index":{"https://app.crunch.io/api/datasets/3/variables/gender/":{"discarded":true}}}'
+            hide_gender
         )
+    })
+    test_that("hideVariables excludes already hidden variables", {
         expect_PATCH(
             ds <- hideVariables(ds, c("gender", "birthyr")),
-            "https://app.crunch.io/api/datasets/3/variables/",
-            '{"element":"shoji:catalog","index":{"https://app.crunch.io/api/datasets/3/variables/gender/":{"discarded":true}}}'
-        ) ## same
-        expect_PATCH(
-            ds <- hideVariables(ds, vg),
-            "https://app.crunch.io/api/datasets/3/variables/",
-            '{"element":"shoji:catalog","index":{"https://app.crunch.io/api/datasets/3/variables/gender/":{"discarded":true}}}'
-        ) ## same
+            hide_gender
+        )
+    })
+    test_that("hideVariables with multiple vars to hide", {
         expect_PATCH(
             ds <- hideVariables(ds, c("gender", "starttime")),
-            "https://app.crunch.io/api/datasets/3/variables/",
-            '{"element":"shoji:catalog","index":{"https://app.crunch.io/api/datasets/3/variables/gender/":{"discarded":true},',
-            '"https://app.crunch.io/api/datasets/3/variables/starttime/":{"discarded":true}}}'
+            hide_both
         )
         expect_PATCH(
             ds <- hideVariables(ds, 1:2),
-            "https://app.crunch.io/api/datasets/3/variables/",
-            '{"element":"shoji:catalog","index":{"https://app.crunch.io/api/datasets/3/variables/gender/":{"discarded":true},',
-            '"https://app.crunch.io/api/datasets/3/variables/starttime/":{"discarded":true}}}'
+            hide_both
         )
     })
 
     test_that("hiddenVariables<- request", {
         expect_PATCH(
             hiddenVariables(ds) <- "gender",
-            "https://app.crunch.io/api/datasets/3/variables/",
-            '{"element":"shoji:catalog","index":{"https://app.crunch.io/api/datasets/3/variables/gender/":{"discarded":true}}}'
+            hide_gender
         )
         expect_PATCH(
             hiddenVariables(ds) <- c("gender", "birthyr"),
-            "https://app.crunch.io/api/datasets/3/variables/",
-            '{"element":"shoji:catalog","index":{"https://app.crunch.io/api/datasets/3/variables/gender/":{"discarded":true}}}'
-        ) ## same
+            hide_gender
+        )
     })
 
+    unhide_birthyr <- paste0(
+        'https://app.crunch.io/api/datasets/3/folders/',
+        ' ',
+        '{',
+            '"element":"shoji:catalog",',
+            '"index":{',
+                '"https://app.crunch.io/api/datasets/3/variables/birthyr/":{}',
+            '}',
+        '}'
+    )
     test_that("unhideVariables with various input types makes the right request", {
         expect_PATCH(
             ds <- unhideVariables(ds, "birthyr"),
-            "https://app.crunch.io/api/datasets/3/variables/",
-            '{"element":"shoji:catalog","index":{"https://app.crunch.io/api/datasets/3/variables/birthyr/":{"discarded":false}}}'
+            unhide_birthyr
         )
         expect_PATCH(
             ds <- unhideVariables(ds, c("gender", "birthyr")),
-            "https://app.crunch.io/api/datasets/3/variables/",
-            '{"element":"shoji:catalog","index":{"https://app.crunch.io/api/datasets/3/variables/birthyr/":{"discarded":false}}}'
-        ) ## same
-        expect_PATCH(
-            ds <- unhideVariables(ds, vg),
-            "https://app.crunch.io/api/datasets/3/variables/",
-            '{"element":"shoji:catalog","index":{"https://app.crunch.io/api/datasets/3/variables/birthyr/":{"discarded":false}}}'
-        ) ## same
+            unhide_birthyr
+        )
         expect_no_request(ds <- unhideVariables(ds, c("gender", "starttime")))
     })
 
     test_that("hide method on variable makes right request", {
         expect_PATCH(
             hide(ds$gender),
-            "https://app.crunch.io/api/datasets/3/variables/",
-            '{"https://app.crunch.io/api/datasets/3/variables/gender/":{"discarded":true}}'
+            hide_gender
         )
         expect_warning(
             expect_no_request(hide(ds$birthyr)),
@@ -128,8 +145,7 @@ with_mock_crunch({
         expect_warning(
             expect_PATCH(
                 unhide(ds$birthyr),
-                "https://app.crunch.io/api/datasets/3/variables/",
-                '{"https://app.crunch.io/api/datasets/3/variables/birthyr/":{"discarded":false}}'
+                unhide_birthyr
             ),
             "Variable birthyr is hidden"
         )