+++ NEWS.md
@@ -1,4 +1,5 @@
 # crunch 1.26.1 (under development)
+* Internally, the `(un)hideVariables()` functions are upgraded to use folder operations.
 
 # crunch 1.26.0